Split Hash Map and driver code.
